New thing learned, the dispatchQueue.global(), the way to send a task to background free the main queue. My problem was: - to free API i use, fackstoreAPI use large image and the downloading of that data took considerable time. - Assigning those image to ui image on the first load make the app crashing because the absence of not downloaded image yet is considered as nil. -blocking the url image download and set to ui image in main queue make the app not crashing but freeze the ui until the download complete and this is bad for user experience.
the discover global() function of dispatchQueue class with async made me separate the process of fetch url image and sent it to background thread to free the main queue, inside the global block i insert the URL image, set the data get into a variable that can be used to set the ui, the ui image assignment is inside a dispatchQueue.main.async which is inside the global scope too.

i'm happy with the result and the app is better on loading.